requirements.txt
actionlib==1.13.2
angles==1.9.13
anyascii==0.2.0
asgiref==3.4.0
beautifulsoup4==4.8.2
bondpy==1.8.6
camera-calibration==1.15.3
camera-calibration-parsers==1.12.0
catkin==0.8.10
certifi==2021.5.30
chardet==4.0.0
controller-manager==0.19.4
controller-manager-msgs==0.19.4
cv-bridge==1.15.0
diagnostic-analysis==1.10.4
diagnostic-common-diagnostics==1.10.4
diagnostic-updater==1.10.4
Django==3.1.12
django-cogwheels==0.3
django-cors-headers==3.7.0
django-filter==2.4.0
django-modelcluster==5.1
django-taggit==1.4.0
django-treebeard==4.5.1
djangorestframework==3.12.4
draftjs-exporter==2.1.7
dynamic-reconfigure==1.7.1
et-xmlfile==1.1.0
gazebo-plugins==2.9.2
gazebo-ros==2.9.2
gencpp==0.6.5
geneus==3.0.0
genlisp==0.4.18
genmsg==0.5.16
gennodejs==2.0.2
genpy==0.6.15
html5lib==1.1
idna==2.10
image-geometry==1.15.0
interactive-markers==1.12.0
joint-state-publisher==1.15.0
joint-state-publisher-gui==1.15.0
l18n==2020.6.1
laser-geometry==1.6.7
message-filters==1.15.11
openpyxl==3.0.7
Pillow==8.2.0
python-qt-binding==0.4.3
pytz==2021.1
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
requests==2.25.1
resource-retriever==1.12.6
rosbag==1.15.11
rosboost-cfg==1.15.7
rosclean==1.15.7
roscreate==1.15.7
rosgraph==1.15.11
roslaunch==1.15.11
roslib==1.15.7
roslint==0.12.0
roslz4==1.15.11
rosmake==1.15.7
rosmaster==1.15.11
rosmsg==1.15.11
rosnode==1.15.11
rosparam==1.15.11
rospy==1.15.11
rosservice==1.15.11
rostest==1.15.11
rostopic==1.15.11
rosunit==1.15.7
roswtf==1.15.11
rqt-action==0.4.9
rqt-bag==0.5.1
rqt-bag-plugins==0.5.1
rqt-console==0.4.11
rqt-dep==0.4.12
rqt-graph==0.4.14
rqt-gui==0.5.2
rqt-gui-py==0.5.2
rqt-image-view==0.4.16
rqt-launch==0.4.9
rqt-logger-level==0.4.11
rqt-moveit==0.5.10
rqt-msg==0.4.10
rqt-nav-view==0.5.7
rqt-plot==0.4.13
rqt-pose-view==0.5.11
rqt-publisher==0.4.10
rqt-py-common==0.5.2
rqt-py-console==0.4.10
rqt-reconfigure==0.5.4
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.13
rqt-robot-steering==0.5.12
rqt-runtime-monitor==0.5.9
rqt-rviz==0.6.1
rqt-service-caller==0.4.10
rqt-shell==0.4.11
rqt-srv==0.4.9
rqt-tf-tree==0.6.2
rqt-top==0.4.10
rqt-topic==0.4.12
rqt-web==0.4.10
rviz==1.14.7
sensor-msgs==1.13.1
six==1.16.0
smach==2.5.0
smach-ros==2.5.0
smclib==1.8.6
soupsieve==2.2.1
sqlparse==0.4.1
tablib==3.0.0
telepath==0.2
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.5
tf2-kdl==0.7.5
tf2-py==0.7.5
tf2-ros==0.7.5
topic-tools==1.15.11
Unidecode==1.2.0
urllib3==1.26.6
wagtail==2.11.8
wagtailmenus==3.0.2
webencodings==0.5.1
Willow==1.4
xacro==1.14.6
xlrd==2.0.1
XlsxWriter==1.4.3
xlwt==1.3.0
